+++ src/messages/Message.ts
@@ -1,6 +1,5 @@
 import { randomUUID } from 'crypto'
 import {
-  type ChatCompletionRequestMessage
 } from 'openai'
 import { mongoClient } from '../mongo/MongoClient'
 import { countTokens } from '../OpenAiHelper'
@@ -9,39 +8,41 @@ export class Message {
   id: string
   guildId: string
   channelId: string
-  userId: string
+  user: string
   content: string
   timestamp: number
-  type: 'system' | 'user' | 'assistant'
+  role: 'system' | 'user' | 'assistant'
   tokens: number
-  chatCompletionRequestMessage: ChatCompletionRequestMessage
   constructor (
     {
       id,
       guildId,
       channelId,
-      userId, content,
-      timestamp, type,
-      chatCompletionRequestMessage
+      user,
+      content,
+      timestamp,
+      role
     }: {
       id?: string
       guildId: string
       channelId: string
-      userId: string
+      user: string
       content: string
       timestamp: number
-      type: 'system' | 'user' | 'assistant'
-      chatCompletionRequestMessage: ChatCompletionRequestMessage }
+      role: 'system' | 'user' | 'assistant' }
   ) {
     this.id = id ?? randomUUID()
     this.guildId = guildId
     this.channelId = channelId
-    this.userId = userId
+    this.user = user
     this.content = content
     this.timestamp = timestamp
-    this.type = type
-    this.chatCompletionRequestMessage = chatCompletionRequestMessage
-    this.tokens = countTokens([chatCompletionRequestMessage])
+    this.role = role
+    this.tokens = countTokens([{
+      role: this.role,
+      content: this.content,
+      name: this.user
+    }])
   }
 
   async save (): Promise<void> {