+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-  Refactored the message schema to be simpler
+
 ## [3.1.4] - 2023-04-10
 
 ### Changed